Updated strings Moved the check for updates to MainPage
