+++ Files/Files.csproj
@@ -164,7 +164,7 @@
     <Compile Include="Converters\StorageDeleteOptionToBooleanConverter.cs" />
     <Compile Include="Converters\StringArrayToString.cs" />
     <Compile Include="Converters\UInt32ToString.cs" />
-    <Compile Include="DataModels\AppUpdater.cs" />
+    <Compile Include="Helpers\AppUpdater.cs" />
     <Compile Include="DataModels\DefaultLanguageModel.cs" />
     <Compile Include="Enums\FileOperationType.cs" />
     <Compile Include="DataModels\SidebarPinnedModel.cs" />
