+++ Files/Strings/fr-FR/Resources.resw
@@ -1098,6 +1098,9 @@
   <data name="PropertyLongitude" xml:space="preserve">
     <value>Longitude</value>
   </data>
+  <data name="PropertyAltitude" xml:space="preserve">
+    <value>Altitude</value>
+  </data>
   <data name="PropertyCameraManufacturer" xml:space="preserve">
     <value>Marque d'appareil</value>
   </data>
@@ -1113,14 +1116,26 @@
   <data name="PropertyAperture" xml:space="preserve">
     <value>Ouverture</value>
   </data>
+  <data name="PropertyFormat" xml:space="preserve">
+    <value>Format</value>
+  </data>
   <data name="PropertyArtist" xml:space="preserve">
     <value>Artiste</value>
   </data>
   <data name="PropertyGenre" xml:space="preserve">
     <value>Genre</value>
   </data>
+  <data name="PropertyDuration" xml:space="preserve">
+    <value>Durée</value>
+  </data>
+  <data name="PropertyProducer" xml:space="preserve">
+    <value>Producteur</value>
+  </data>
   <data name="PropertyYear" xml:space="preserve">
     <value>Année</value>
+  </data>
+  <data name="PropertyVersion" xml:space="preserve">
+    <value>Version</value>
   </data>
   <data name="PropertySectionVideo" xml:space="preserve">
     <value>Vidéo</value>
@@ -1140,6 +1155,9 @@
   <data name="PropertySaveErrorDialog.CloseButtonText" xml:space="preserve">
     <value>Annuler</value>
   </data>
+  <data name="ClearPropertiesFlyoutText.Text" xml:space="preserve">
+    <value>Certaines propriétés peuvent contenir des informations personnelles.</value>
+  </data>
   <data name="ClearPropertiesButton.Content" xml:space="preserve">
     <value>Supprimer les propriétés</value>
   </data>
@@ -1148,5 +1166,20 @@
   </data>
   <data name="PropertyColorSpace_Value65535" xml:space="preserve">
     <value>Non spécifié</value>
+  </data>
+  <data name="SettingsAboutLicense.AutomationProperties.Name" xml:space="preserve">
+    <value>Licence</value>
+  </data>
+  <data name="SettingsAboutLicense.Content" xml:space="preserve">
+    <value>Licence</value>
+  </data>
+  <data name="SettingsAboutWebsite.AutomationProperties.Name" xml:space="preserve">
+    <value>Site web</value>
+  </data>
+  <data name="SettingsAboutWebsite.Content" xml:space="preserve">
+    <value>Site web</value>
+  </data>
+  <data name="SettingsAppearanceDateFormatsTip.AutomationProperties.Name" xml:space="preserve">
+    <value>En savoir plus sur les formats de date</value>
   </data>
 </root>
