+++ Files/Views/MainPage.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using Files.Controllers;
 using Files.Controls;
+using Files.Helpers;
 using Files.Filesystem;
 using Files.UserControls.MultitaskingControl;
 using Files.View_Models;
@@ -160,6 +161,18 @@ namespace Files.Views
                 {
                     await AddNewTabByPathAsync(typeof(ModernShellPage), navArgs);
                 }
+
+                // Check for required updates
+                try
+                {
+                    AppUpdater updater = new AppUpdater();
+                    await updater.CheckForUpdatesAsync();
+                }
+                catch (Exception)
+                {
+                    // App is not installed from the store or checking for updates failed
+                }
+                
 
                 // Initial setting of SelectedTabItem
                 Frame rootFrame = Window.Current.Content as Frame;
